[SC] TokenStandard adjustments (#2871)
* Replace uint with ulong

* Rename Transfer method

* Add approval check for previous allowance

* Add xml doc

* Rearrange files and remove duplicate interface

* Test fix

+++ src/Stratis.SmartContracts.IntegrationTests/ContractInternalTransferTests.cs
@@ -628,7 +628,7 @@ namespace Stratis.SmartContracts.IntegrationTests
         [Fact]
         public void Token_Standards_Test()
         {
-            const uint totalSupply = 100_000;
+            const ulong totalSupply = 100_000;
             // Deploy contract
             ContractCompilationResult compilationResult = ContractCompiler.CompileFile("SmartContracts/StandardToken.cs");
             Assert.True(compilationResult.Success);
